@@ -62,58 +61,63 @@
     "underscore": "^1.9.1"
   },
   "devDependencies": {
+    "@babel/core": "^7.8.7",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
     "@babel/preset-env": "^7.8.7",
-    "babel-core": "^6.17.0",
-    "babel-eslint": "^7.2.3",
-    "babel-jest": "^25.1.0",
-    "babel-loader": "^7.0.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-react-hmre": "^1.1.1",
-    "babel-preset-stage-1": "^6.24.1",
-    "base-href-webpack-plugin": "^1.0.2",
+    "@babel/preset-es2015": "^7.0.0-beta.53",
+    "@babel/preset-react": "^7.8.3",
+    "@fortawesome/fontawesome-svg-core": "^1.2.27",
+    "@fortawesome/free-brands-svg-icons": "^5.12.1",
+    "@fortawesome/free-regular-svg-icons": "^5.12.1",
+    "@fortawesome/free-solid-svg-icons": "^5.12.1",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.0.6",
+    "base-href-webpack-plugin": "^2.0.0",
     "bootstrap-webpack": "0.0.x",
-    "copy-webpack-plugin": "^4.1.1",
-    "css-loader": "^0.25.0",
+    "copy-webpack-plugin": "^5.1.1",
+    "css-loader": "^3.4.2",
     "electron": "^1.7.9",
     "enzyme": "3",
     "enzyme-adapter-react-16": "^1.14.0",
-    "eslint": "^4.0.0",
-    "eslint-loader": "^1.8.0",
+    "eslint": "^6.8.0",
+    "eslint-loader": "^3.0.3",
     "eslint-plugin-react": "^7.1.0",
-    "exports-loader": "^0.6.3",
+    "exports-loader": "^0.7.0",
     "expose-loader": "^0.7.1",
-    "extract-text-webpack-plugin": "^2.0.0",
-    "file-loader": "^0.9.0",
-    "font-awesome-webpack": "0.0.4",
-    "html-webpack-plugin": "^2.30.1",
+    "file-loader": "^5.1.0",
+    "font-awesome-webpack": "^0.0.5-beta.2",
+    "font-awesome-webpack-4": "^1.0.0",
+    "html-webpack-plugin": "^4.0.0-alpha.2",
     "husky": "^0.14.3",
-    "imports-loader": "^0.6.5",
+    "imports-loader": "^0.8.0",
+    "interpolate-html-plugin": "^3.0.0",
     "jest": "^19.0.2",
     "less": "^2.7.2",
-    "less-loader": "^4.0.4",
+    "mini-css-extract-plugin": "^0.9.0",
     "node-sass": "^4.8.3",
     "npm-version": "^1.1.0",
+    "preload-webpack-plugin": "3.0.0-beta.3",
     "prettier": "1.18.1",
     "pretty-quick": "^2.0.1",
     "react-addons-test-utils": "^15.6.0",
     "react-lazy-render": "0.3.0",
     "react-test-renderer": "16",
-    "sass-loader": "^6.0.7",
+    "sass": "^1.26.2",
+    "sass-loader": "^8.0.2",
     "scss-loader": "^0.0.1",
-    "style-loader": "^0.13.1",
-    "ts-jest": "^25.2.1",
+    "style-loader": "^1.1.3",
     "uglify-js": "^3.3.7",
-    "url-loader": "^0.5.7",
-    "webpack": "^2.0.0",
-    "webpack-dev-server": "^2.9.7",
+    "url-loader": "^3.0.0",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3",
     "webpack-eslint-plugin": "^1.2.0"
   },
   "scripts": {
     "build": "webpack",
     "build:library": "webpack --config webpack.config.library.js",
     "lint": "eslint --ext .js,.jsx --ignore-pattern public,static,dist ./src/",
-    "jest": "node ./node_modules/jest-cli/bin/jest.js",
+    "test": "node ./node_modules/jest-cli/bin/jest.js",
     "start": "node_modules/.bin/webpack-dev-server --open",
     "electron": "webpack && node_modules/.bin/electron electron.js",
     "format": "prettier 'src/**/*.{js,jsx}' --write",
