@@ -33,4 +33,4 @@ class ErrorBoundary extends React.Component {
   }
 }
 
-module.exports.ErrorBoundary = ErrorBoundary;
+export { ErrorBoundary };
