@@ -1145,9 +1145,11 @@ var DatamonkeyTimersTable = createReactClass({
   }
 });
 
-module.exports.DatamonkeyTable = DatamonkeyTable;
-module.exports.DatamonkeyTableRow = DatamonkeyTableRow;
-module.exports.DatamonkeyRateDistributionTable = DatamonkeyRateDistributionTable;
-module.exports.DatamonkeyPartitionTable = DatamonkeyPartitionTable;
-module.exports.DatamonkeyModelTable = DatamonkeyModelTable;
-module.exports.DatamonkeyTimersTable = DatamonkeyTimersTable;
+export {
+  DatamonkeyTable,
+  DatamonkeyTableRow,
+  DatamonkeyRateDistributionTable,
+  DatamonkeyPartitionTable,
+  DatamonkeyModelTable,
+  DatamonkeyTimersTable
+};
