@@ -58,7 +58,8 @@ function FUBARSummary(props) {
               min="0"
               max="1"
               onChange={props.updatePosteriorProbability}
-            />.
+            />
+            .
           </p>
           <hr />
           <p>
@@ -68,7 +69,8 @@ function FUBARSummary(props) {
                 here
               </a>{" "}
               for more information about the FUBAR method.
-              <br />Please cite{" "}
+              <br />
+              Please cite{" "}
               <a
                 href="http://www.ncbi.nlm.nih.gov/pubmed/23420840"
                 id="summary-pmid"
@@ -180,9 +182,10 @@ class FUBARViz extends React.Component {
       .attr("class", "axis axis")
       .call(yAxis);
 
-    d3
-      .selectAll(".x-axis > .tick > text")
-      .attr("transform", "rotate(-90) translate(-20, -15)");
+    d3.selectAll(".x-axis > .tick > text").attr(
+      "transform",
+      "rotate(-90) translate(-20, -15)"
+    );
 
     main
       .append("text")
@@ -290,8 +293,7 @@ class FUBARViz extends React.Component {
         .domain([0, d3.max(data.map(row => +row[2]))])
         .range([0, this.width / this.n_gridpoints]);
 
-      d3
-        .selectAll("circle")
+      d3.selectAll("circle")
         .data(data)
         .attr("cx", d => this.x(+d[0].toFixed(2)))
         .attr("cy", d => this.y(+d[1].toFixed(2)))
@@ -330,8 +332,9 @@ class FUBARViz extends React.Component {
           />
           {self.state.input_error ? (
             <div className="alert alert-danger">
-              <span className="fas fa-exclamation" aria-hidden="true" />Enter a
-              valid site (a number from 1 to {this.props.number_of_sites}).
+              <span className="fas fa-exclamation" aria-hidden="true" />
+              Enter a valid site (a number from 1 to{" "}
+              {this.props.number_of_sites}).
             </div>
           ) : (
             ""
@@ -638,5 +641,4 @@ function render_fubar(data, element, fasta, originalFile, analysisLog) {
   );
 }
 
-module.exports = render_fubar;
-module.exports.FUBAR = FUBAR;
+export { FUBAR, render_fubar };
