@@ -38,17 +38,14 @@ function GARDResults(props) {
     .reduce((a, b) => a + b, 0);
   var totalModelCount = props.data.models.length;
   var percentageExplored = (
-    100 *
-    totalModelCount /
+    (100 * totalModelCount) /
     totalPossibleModels
   ).toFixed(2);
   var evidence_statement = props.data.improvements ? (
     <span>
       <strong className="hyphy-highlight">found evidence</strong> of{" "}
-      {props.data.lastImprovedBPC} recombination breakpoint{props.data
-        .lastImprovedBPC == 1
-        ? ""
-        : "s"}
+      {props.data.lastImprovedBPC} recombination breakpoint
+      {props.data.lastImprovedBPC == 1 ? "" : "s"}
     </span>
   ) : (
     <span>
@@ -78,7 +75,8 @@ function GARDResults(props) {
                 here
               </a>{" "}
               for more information about this method.
-              <br />Please cite{" "}
+              <br />
+              Please cite{" "}
               <a
                 href="http://www.ncbi.nlm.nih.gov/pubmed/16818476"
                 id="summary-pmid"
@@ -138,7 +136,7 @@ function GARDRecombinationReport(props) {
             />
             <text
               x={scale(bp_delta / 2)}
-              y={2 * height / 3}
+              y={(2 * height) / 3}
               fill={"white"}
               textAnchor={"middle"}
             >
@@ -368,14 +366,14 @@ function GARDTopologyReport(props) {
         </table>
         <p className="description">
           Comparing the AIC<sub>c</sub> score of the best fitting GARD model,
-          that allows for different topologies between segments ({currentAIC.toFixed(
-            1
-          )}), and that of the model that assumes the same tree for all the
-          partitions inferred by GARD the same tree, but allows different branch
-          lengths between partitions ({props.data.singleTreeAICc.toFixed(1)})
-          suggests that because the multiple tree model{" "}
-          {w > 0.01 ? "cannot" : "can"} be preferred over the single tree model
-          by an evidence ratio of 100 or greater, {conclusion}.
+          that allows for different topologies between segments (
+          {currentAIC.toFixed(1)}), and that of the model that assumes the same
+          tree for all the partitions inferred by GARD the same tree, but allows
+          different branch lengths between partitions (
+          {props.data.singleTreeAICc.toFixed(1)}) suggests that because the
+          multiple tree model {w > 0.01 ? "cannot" : "can"} be preferred over
+          the single tree model by an evidence ratio of 100 or greater,{" "}
+          {conclusion}.
         </p>
         <p className="description">
           Please consult the above Kishino Hasegawa topological incongruence
@@ -457,4 +455,4 @@ class GARD_HyPhy_2_3 extends React.Component {
   }
 }
 
-module.exports.GARD_HyPhy_2_3 = GARD_HyPhy_2_3;
+export { GARD_HyPhy_2_3 };
